+++ docs/changelog.rst
@@ -14,7 +14,7 @@ version **0.2.2** (2012 ??? ??)
 *******************************
 
 * add ``wait_for_and_refresh`` action
-
+* ``set_dropdown_value`` can set text or value now 
 
 version **0.2.1** (2012 Apr 22)
 *******************************